+++ CHANGELOG.md
@@ -5,7 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.6.0] - 2021-06-23
+## [0.8.0] - 2021-06-25
+
+### 🔥 Added
+
+- `edit RFQ` page
+- `delete RFQ` page
+- sepatate admin view
+- front page
+- styling
+
+## [0.7.0] - 2021-06-25
+
+### 🔥 Added
+
+- `delete RFQ's requirement` component
+
+## [0.6.0] - 2021-06-24
 
 ### 🔥 Added